# src/api/wwwroot/openapi.yaml
openapi: 3.0.0
info:
  description: Restoplan API
  version: 3.0.0
  title: Restoplan API
  contact:
    email: azdevteam@microsoft.com

components:
  schemas:
    RestoplanItem:
      type: object
      required:
        - listId
        - name
        - description
      description: A task that needs to be completed
      properties:
        id:
          type: string
        listId:
          type: string
        name:
          type: string
        description:
          type: string
        state:
          $ref: "#/components/schemas/RestoplanState"
        dueDate:
          type: string
          format: date-time
        completedDate:
          type: string
          format: date-time
        startDate:
          type: string
          format: date-time
    RestoplanList:
      type: object
      required:
        - name
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
      description: " A list of related Restoplan items"
    RestoplanState:
      type: string
      enum:
        - todo
        - inprogress
        - done
  parameters:
    listId:
      in: path
      required: true
      name: listId
      description: The Restoplan list unique identifier
      schema:
        type: string
    itemId:
      in: path
      required: true
      name: itemId
      description: The Restoplan item unique identifier
      schema:
        type: string
    state:
      in: path
      required: true
      name: state
      description: The Restoplan item state
      schema:
        $ref: "#/components/schemas/RestoplanState"
    top:
      in: query
      required: false
      name: top
      description: The max number of items to returns in a result
      schema:
        type: number
        default: 20
    skip:
      in: query
      required: false
      name: skip
      description: The number of items to skip within the results
      schema:
        type: number
        default: 0

  requestBodies:
    RestoplanList:
      description: The Restoplan List
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RestoplanList"
    RestoplanItem:
      description: The Restoplan Item
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RestoplanItem"

  responses:
    RestoplanList:
      description: A Restoplan list result
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RestoplanList"
    TodoListArray:
      description: An array of Restoplan lists
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/RestoplanList"
    RestoplanItem:
      description: A Restoplan item result
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RestoplanItem"
    TodoItemArray:
      description: An array of Restoplan items
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/RestoplanItem"

paths:
  /lists:
    get:
      operationId: GetLists
      summary: Gets An array of Restoplan lists
      tags:
        - Lists
      parameters:
        - $ref: "#/components/parameters/top"
        - $ref: "#/components/parameters/skip"
      responses:
        200:
          $ref: "#/components/responses/TodoListArray"
    post:
      operationId: CreateList
      summary: Creates a new Restoplan list
      tags:
        - Lists
      requestBody:
        $ref: "#/components/requestBodies/RestoplanList"
      responses:
        201:
          $ref: "#/components/responses/RestoplanList"
        400:
          description: Invalid request schema
  /lists/{listId}:
    get:
      operationId: GetListById
      summary: Gets a Restoplan list by unique identifier
      tags:
        - Lists
      parameters:
        - $ref: "#/components/parameters/listId"
      responses:
        200:
          $ref: "#/components/responses/RestoplanList"
        404:
          description: Restoplan list not found
    put:
      operationId: UpdateListById
      summary: Updates a Restoplan list by unique identifier
      tags:
        - Lists
      requestBody:
        $ref: "#/components/requestBodies/RestoplanList"
      parameters:
        - $ref: "#/components/parameters/listId"
      responses:
        200:
          $ref: "#/components/responses/RestoplanList"
        404:
          description: Restoplan list not found
        400:
          description: Restoplan list is invalid
    delete:
      operationId: DeleteListById
      summary: Deletes a Restoplan list by unique identifier
      tags:
        - Lists
      parameters:
        - $ref: "#/components/parameters/listId"
      responses:
        204:
          description: Restoplan list deleted successfully
        404:
          description: Restoplan list not found
  /lists/{listId}/items:
    post:
      operationId: CreateItem
      summary: Creates a new Restoplan item within a list
      tags:
        - Items
      requestBody:
        $ref: "#/components/requestBodies/RestoplanItem"
      parameters:
        - $ref: "#/components/parameters/listId"
      responses:
        201:
          $ref: "#/components/responses/RestoplanItem"
        404:
          description: Restoplan list not found
    get:
      operationId: GetItemsByListId
      summary: Gets Restoplan items within the specified list
      tags:
        - Items
      parameters:
        - $ref: "#/components/parameters/listId"
        - $ref: "#/components/parameters/top"
        - $ref: "#/components/parameters/skip"
      responses:
        200:
          $ref: "#/components/responses/TodoItemArray"
        404:
          description: Restoplan list not found
  /lists/{listId}/items/{itemId}:
    get:
      operationId: GetItemById
      summary: Gets a Restoplan item by unique identifier
      tags:
        - Items
      parameters:
        - $ref: "#/components/parameters/listId"
        - $ref: "#/components/parameters/itemId"
      responses:
        200:
          $ref: "#/components/responses/RestoplanItem"
        404:
          description: Restoplan list or item not found
    put:
      operationId: UpdateItemById
      summary: Updates a Restoplan item by unique identifier
      tags:
        - Items
      requestBody:
        $ref: "#/components/requestBodies/RestoplanItem"
      parameters:
        - $ref: "#/components/parameters/listId"
        - $ref: "#/components/parameters/itemId"
      responses:
        200:
          $ref: "#/components/responses/RestoplanItem"
        400:
          description: Restoplan item is invalid
        404:
          description: Restoplan list or item not found
    delete:
      operationId: DeleteItemById
      summary: Deletes a Restoplan item by unique identifier
      tags:
        - Items
      parameters:
        - $ref: "#/components/parameters/listId"
        - $ref: "#/components/parameters/itemId"
      responses:
        204:
          description: Restoplan item deleted successfully
        404:
          description: Restoplan list or item not found
  /lists/{listId}/items/state/{state}:
    get:
      operationId: GetItemsByListIdAndState
      summary: Gets a list of Restoplan items of a specific state
      tags:
        - Items
      parameters:
        - $ref: "#/components/parameters/listId"
        - $ref: "#/components/parameters/state"
        - $ref: "#/components/parameters/top"
        - $ref: "#/components/parameters/skip"
      responses:
        200:
          $ref: "#/components/responses/TodoItemArray"
        404:
          description: Restoplan list or item not found
    put:
      operationId: UpdateItemsStateByListId
      summary: Changes the state of the specified list items
      tags:
        - Items
      requestBody:
        description: unique identifiers of the Restoplan items to update
        content:
          application/json:
            schema:
              type: array
              items:
                description: The Restoplan item unique identifier
                type: string
      parameters:
        - $ref: "#/components/parameters/listId"
        - $ref: "#/components/parameters/state"
      responses:
        204:
          description: Restoplan items updated
        400:
          description: Update request is invalid